requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.0
Pillow==10.4.0
PyYAML==6.0.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
streamlit==1.37.1
